Version 0.1.4: Add SIMD and Rayon support
Using the new SIMD iterators in `imgref-iter`, it's possible to
implement SIMD version of stackblur. Right now SIMD doesn't offer too
much of a benefit, but it is still a benefit nonetheless, and rayon
definitely offers a huge benefit.

This has been sitting around in PR hell for far too long, and I don't
like managing multiple disparate branches, so it's about time this gets
merged and released.