Version 0.1.6: fix UB in blur methods
Obtaining the column reader before writing to the rows is UB because
writing to the rows invalidates the column reader's borrow.

This release fixes that. It does not update imgref-iter yet because that
will be a breaking change (due to the huge MSRV hike). The version of
imgref-iter used here does not pass Miri yet (the latest version does).